+++ packages/command-tests/commands.js
@@ -14,12 +14,6 @@ const LIMIT_TESTS = !!process.env.LIMIT_TESTS;
 const PARALLELISM = 4;
 
 const CONFIGS = [
-    ["web", "full", "ts", "8.0"],
-    ["native", "full", "ts", "8.6"],
-    ["web", "full", "ts", "8.6"],
-    ["web", "full", "js", "8.7"],
-    ["web", "full", "ts", "8.9"],
-    ["native", "full", "ts", "8.9"],
     ["web", "full", "js", "latest"],
     ["web", "full", "ts", "latest"],
     ["native", "full", "js", "latest"],
@@ -376,14 +370,15 @@ async function execFailedAsync(command, workDir) {
 function fixPackageJson(json) {
     const devDependencies = {
         "@types/jest": "^29.0.0",
-        "@types/react": "~18.2.0",
-        "@types/react-native": "~0.72.0",
-        "@types/react-dom": "~18.2.0",
+        "@types/react": "^19.0.0",
+        "@types/react-native": "0.78.2",
+        "@types/react-dom": "^19.0.0",
         "@types/react-test-renderer": "~18.0.0"
     };
     const overrides = {
-        react: "18.2.0",
-        "react-native": "0.72.7"
+        "react": "^19.0.0",
+        "react-dom": "^19.0.0",
+        "react-native": "0.78.2"
     };
 
     Object.keys(devDependencies)