+++ packages/pluggable-widgets-tools/utils/migration.js
@@ -23,7 +23,7 @@ const dependencies = [
     { name: "@types/react-native", version: "remove", check: CheckType.MINOR },
     { name: "@types/react-native-push-notification", version: "8.1.1", check: CheckType.MAJOR_MINOR },
     { name: "@types/react-dom", version: "remove", check: CheckType.MAJOR },
-    { name: "@types/react-test-renderer", version: "18.0.0", check: CheckType.MAJOR },
+    { name: "@types/react-test-renderer", version: "19.0.0", check: CheckType.MAJOR },
     { name: "@types/enzyme-adapter-react-16", version: "remove", check: CheckType.MAJOR },
     { name: "@react-native-firebase/app", version: "17.3.0", check: CheckType.MAJOR_MINOR },
     { name: "@react-native-firebase/messaging", version: "17.3.0", check: CheckType.MAJOR_MINOR },
@@ -46,9 +46,9 @@ const dependencies = [
     { name: "react-native-webview", version: "11.26.1", check: CheckType.MAJOR_MINOR }
 ];
 
-const reactPackage = { version: "18.2.0", check: CheckType.MAJOR_MINOR };
-const reactDomPackage = { version: "18.2.0", check: CheckType.MAJOR_MINOR };
-const reactNativePackage = { version: "0.72.7", check: CheckType.MINOR };
+const reactPackage = { version: "19.0.0", check: CheckType.MAJOR_MINOR };
+const reactDomPackage = { version: "19.0.0", check: CheckType.MAJOR_MINOR };
+const reactNativePackage = { version: "0.78.2", check: CheckType.MINOR };
 
 const resolutionsOverrides = [
     { name: "react", ...reactPackage },
@@ -226,6 +226,7 @@ async function checkMigration() {
                     execSync("shx rm -rf ./{node_modules,package-lock.json}", { cwd: process.cwd(), stdio: "inherit" });
                     console.log("Done.");
                     execSync(`npm install`, { cwd: process.cwd(), stdio: "inherit" });
+                    execSync(`npx eslint --no-eslintrc --rule '@typescript-eslint/no-unused-vars: ["error", { enableAutofixRemoval: { imports: true } }]' --ext ts,tsx,js,jsx --parser @typescript-eslint/parser --plugin '@typescript-eslint' --fix ./src`, { cwd: process.cwd(), stdio: "inherit" })
                 } catch (e) {
                     console.log(red("An error occurred while auto updating your dependencies"));
                     console.error(e);